Make IngestRecordsITBase extend IngestRecordsTestBase
