+++ java/ingest/src/test/java/sleeper/ingest/IngestRecordsITBase.java
@@ -21,54 +21,24 @@ import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
-import org.junit.Before;
 import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.rules.TemporaryFolder;
 import org.testcontainers.containers.GenericContainer;
-import sleeper.configuration.properties.InstanceProperties;
-import sleeper.configuration.properties.table.TableProperties;
 import sleeper.core.CommonTestConstants;
-import sleeper.core.record.Record;
-import sleeper.core.schema.Field;
 import sleeper.core.schema.Schema;
-import sleeper.core.schema.type.LongType;
-import sleeper.ingest.testutils.IngestRecordsTestDataHelper;
-import sleeper.statestore.FixedStateStoreProvider;
-import sleeper.statestore.StateStore;
 import sleeper.statestore.StateStoreException;
 import sleeper.statestore.dynamodb.DynamoDBStateStore;
 import sleeper.statestore.dynamodb.DynamoDBStateStoreCreator;
 
-import java.util.List;
 import java.util.UUID;
 
-import static sleeper.ingest.testutils.IngestRecordsTestDataHelper.TEST_TABLE_NAME;
-import static sleeper.ingest.testutils.IngestRecordsTestDataHelper.defaultInstanceProperties;
-import static sleeper.ingest.testutils.IngestRecordsTestDataHelper.defaultTableProperties;
-import static sleeper.ingest.testutils.IngestRecordsTestDataHelper.schemaWithRowKeys;
-
-public class IngestRecordsITBase {
+public class IngestRecordsITBase extends IngestRecordsTestBase {
     private static final int DYNAMO_PORT = 8000;
 
     @ClassRule
     public static GenericContainer dynamoDb = new GenericContainer(CommonTestConstants.DYNAMODB_LOCAL_CONTAINER)
             .withExposedPorts(DYNAMO_PORT);
-    @Rule
-    public TemporaryFolder folder = new TemporaryFolder(CommonTestConstants.TMP_DIRECTORY);
 
-    protected final Field field = new Field("key", new LongType());
-    protected final Schema schema = schemaWithRowKeys(field);
-    protected String inputFolderName;
-    protected String sketchFolderName;
-
-    @Before
-    public void setUpBase() throws Exception {
-        inputFolderName = folder.newFolder().getAbsolutePath();
-        sketchFolderName = folder.newFolder().getAbsolutePath();
-    }
-
-    protected static DynamoDBStateStore getStateStore(Schema schema)
+    protected DynamoDBStateStore getStateStore(Schema schema)
             throws StateStoreException {
         AwsClientBuilder.EndpointConfiguration endpointConfiguration =
                 new AwsClientBuilder.EndpointConfiguration("http://" + dynamoDb.getContainerIpAddress() + ":"
@@ -82,16 +52,5 @@ public class IngestRecordsITBase {
         DynamoDBStateStore stateStore = dynamoDBStateStoreCreator.create();
         stateStore.initialise();
         return stateStore;
-    }
-
-    protected IngestFactory createIngestFactory(StateStore stateStore, TableProperties tableProperties, InstanceProperties instanceProperties) {
-        return IngestRecordsTestDataHelper.createIngestFactory(inputFolderName, new FixedStateStoreProvider(tableProperties, stateStore), instanceProperties);
-    }
-
-    protected IngestResult ingestRecords(Schema schema, StateStore stateStore, List<Record> records) {
-        InstanceProperties instanceProperties = defaultInstanceProperties();
-        TableProperties tableProperties = defaultTableProperties(schema, TEST_TABLE_NAME, sketchFolderName, instanceProperties);
-        IngestFactory factory = createIngestFactory(stateStore, tableProperties, instanceProperties);
-        return factory.ingestRecords(tableProperties, records);
     }
 }
